refactor cjs utils into @nuxt/utils
