+++ packages/cli/src/utils/config.js
@@ -2,6 +2,7 @@ import path from 'path'
 import consola from 'consola'
 import defaultsDeep from 'lodash/defaultsDeep'
 import { defaultNuxtConfigFile, getDefaultNuxtConfig } from '@nuxt/config'
+import { clearRequireCache, scanRequireTree } from '@nuxt/utils'
 import esm from 'esm'
 
 export async function loadNuxtConfig(argv) {
@@ -66,39 +67,4 @@ export async function loadNuxtConfig(argv) {
   }, options.server || {}, getDefaultNuxtConfig().server)
 
   return options
-}
-
-function clearRequireCache(id) {
-  const entry = require.cache[id]
-  if (!entry || id.includes('node_modules')) {
-    return
-  }
-
-  if (entry.parent) {
-    const i = entry.parent.children.findIndex(e => e.id === id)
-    if (i > -1) {
-      entry.parent.children.splice(i, 1)
-    }
-  }
-
-  for (const child of entry.children) {
-    clearRequireCache(child.id)
-  }
-
-  delete require.cache[id]
-}
-
-function scanRequireTree(id, files = new Set()) {
-  const entry = require.cache[id]
-  if (!entry || id.includes('node_modules') || files.has(id)) {
-    return files
-  }
-
-  files.add(entry.id)
-
-  for (const child of entry.children) {
-    scanRequireTree(child.id, files)
-  }
-
-  return files
 }
