+++ packages/utils/src/index.js
@@ -6,3 +6,4 @@ export * from './route'
 export * from './serialize'
 export * from './task'
 export * from './timer'
+export * from './cjs'
